refactor: phase 2 media visual layers, playback store, fullscreen portal
